Integration txtorcon torDeferredDispatcher
When the tor client endpoint `listen` method is
called and the txtorcon module is installed then
we utilized the torDeferredDispatcher to receive
the socks_port of the previously started tor proc.
We add this socks_port to the beginning of our list
of socks ports; we attempt a connection to each one
until a connection succeeds or we consume the list and fail.